Fix Room Api

+++ apps/backend/src/app/services/rooms.service.ts
@@ -1,14 +1,9 @@
 import { BadRequestException, Injectable, Logger } from '@nestjs/common';
 import { AddRoomRequest, UpdateRoomRequest } from '@app/models';
-import { UpdateResult } from 'typeorm';
 import { Rooms } from '../models';
 import { RoomsRepository } from '../repositories';
-import { RoomsRequestPayload } from '../payload/request/rooms.payload';
-import { RoomsResponsePayload } from '../payload/response/rooms.payload';
 import { KeycloakUserInstance } from '../dto/keycloak.user';
-import { Direction } from '../models/search-pagination.payload';
 import { ChooseBookingRoomFilterPayload } from '../payload/request/choose-booking-room-filter.payload';
-import { IPaginationMeta, Pagination } from 'nestjs-typeorm-paginate';
 import { RoomsPaginationParams } from '../controllers/rooms-pagination.model';
 
 @Injectable()
@@ -24,7 +19,7 @@ export class RoomsService {
         throw new BadRequestException('This room is already existed');
       }
 
-      const addedRoom = await this.repository.save(
+      return await this.repository.save(
         {
           createdBy: user.account_id,
           disabledBy: user.account_id,
@@ -35,8 +30,6 @@ export class RoomsService {
           transaction: true,
         }
       );
-
-      return addedRoom;
     } catch (e) {
       this.logger.error(e.message);
       throw new BadRequestException(
@@ -81,8 +74,7 @@ export class RoomsService {
 
   async getDisabledRooms(search: string): Promise<Rooms[]> {
     try {
-      const data = await this.repository.findDisabledRooms(search);
-      return data;
+      return await this.repository.findDisabledRooms(search);
     } catch (e) {
       this.logger.error(e);
       throw new BadRequestException(
@@ -121,7 +113,13 @@ export class RoomsService {
     body: UpdateRoomRequest
   ): Promise<void> {
     let room;
-
+    if (
+      body.name.length < 1 ||
+      body.type.length < 1 ||
+      body.description.length < 1
+    ) {
+      throw new BadRequestException('Fields cannot be left blank!');
+    }
     try {
       room = await this.repository.findOneOrFail({
         where: {
@@ -132,6 +130,7 @@ export class RoomsService {
       this.logger.error(e.message);
       throw new BadRequestException("Room doesn't exist with the provided id");
     }
+
     if (room.name !== body.name) {
       const isExisted = await this.repository.isExistedByName(body.name);
       if (isExisted) {
@@ -162,6 +161,10 @@ export class RoomsService {
 
   async disableById(accountId: string, id: string): Promise<any> {
     try {
+      const rooms = await this.repository.findDisabledRooms('');
+      if (rooms.findIndex((room) => room.id === id) !== -1) {
+        throw new BadRequestException('Room already disable!');
+      }
       const result = await this.repository.disableById(accountId, id);
       if (result.affected < 1) {
         throw new BadRequestException(
@@ -176,6 +179,19 @@ export class RoomsService {
 
   async handleRestoreDeletedRoomById(id: string) {
     try {
+      const room = await this.repository.findOneOrFail({
+        where: {
+          id: id,
+        },
+      });
+      if (
+        room.deletedBy == null &&
+        room.deletedAt == null &&
+        room.disabledBy == null &&
+        room.disabledAt == null
+      ) {
+        throw new BadRequestException('This room is active!');
+      }
       const result = await this.repository.restoreDeletedRoomById(id);
       if (result.affected < 1) {
         throw new BadRequestException(
@@ -192,6 +208,19 @@ export class RoomsService {
 
   async handleRestoreDisabledRoomById(id: string) {
     try {
+      const room = await this.repository.findOneOrFail({
+        where: {
+          id: id,
+        },
+      });
+      if (
+        room.disabledBy == null &&
+        room.disabledAt == null &&
+        room.disabledBy == null &&
+        room.disabledAt == null
+      ) {
+        throw new BadRequestException('This room is active!');
+      }
       const result = await this.repository.restoreDisabledRoomById(id);
       if (result.affected < 1) {
         throw new BadRequestException(
@@ -208,6 +237,10 @@ export class RoomsService {
 
   async deleteById(accountId: string, id: string) {
     try {
+      const rooms = await this.repository.findDisabledRooms('');
+      if (rooms.findIndex((room) => room.id === id) !== -1) {
+        throw new BadRequestException('Room already deleted!');
+      }
       const result = await this.repository.deleteById(accountId, id);
       if (result.affected < 1) {
         throw new BadRequestException(
